# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeMarket API",
    "version": "1.3.0",
    "description": "WeMarket QR Menu & Small Business Platform API"
  },
  "servers": [
    {
      "url": "https://wemarket.onrender.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "orderCapability": {
        "type": "apiKey",
        "in": "header",
        "name": "x-order-capability"
      }
    }
  },
...
  "paths": {
...
    "/api/analytics/store/{storeId}/products": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "인기 판매 상품 순위 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "인기 상품 순위 목록"
          }
        }
      }
    },
...
    "/api/inventory/products/{productId}/stock": {
      "put": {
        "tags": [
          "Inventory"
        ],
        "summary": "상품별 재고 수동 조정 (입출고)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "change",
                  "reason"
                ],
                "properties": {
                  "change": {
                    "type": "integer",
                    "description": "양수=입고, 음수=차감"
                  },
                  "reason": {
                    "type": "string",
                    "enum": [
                      "MANUAL_IN",
                      "MANUAL_OUT",
                      "CORRECTION",
                      "RETURN"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "재고 조정 완료"
          }
        }
      }
    },
...
    "/api/products/store/{storeId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "매장별 상품 목록 조회 (메뉴판)",
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상품 목록 (60초 캐시)"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "상품 상세 조회",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상품 상세 정보"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "상품 정보 수정",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 완료"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "상품 삭제",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 완료"
          }
        }
      }
    },
    "/api/products": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "상품 생성",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "상품 생성 완료"
          }
        }
      }
    },
    "/api/products/bulk": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "상품 일괄 등록",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "store_id",
                  "products"
                ],
                "properties": {
                  "store_id": {
                    "type": "integer"
                  },
                  "products": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "일괄 등록 완료"
          }
        }
      }
    },
    "/api/products/import": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "다른 매장에서 메뉴 가져오기",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "source_store_id",
                  "target_store_id"
                ],
                "properties": {
                  "source_store_id": {
                    "type": "integer"
                  },
                  "target_store_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "가져오기 완료"
          }
        }
      }
    },
    "/api/products/{id}/status": {
      "patch": {
        "tags": [
          "Products"
        ],
        "summary": "상품 상태 변경",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "상태 변경 완료"
          }
        }
      }
    },
    "/api/products/{id}/options": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "상품 옵션 생성",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductOptionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "옵션 생성 완료"
          }
        }
      }
    },
    "/api/products/{id}/options/{optionId}": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "상품 옵션 수정",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "optionId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductOptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "옵션 수정 완료"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "상품 옵션 삭제",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "optionId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "옵션 삭제 완료"
          }
        }
      }
    },
    "/api/products/{id}/option-items": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "상품 옵션 항목 생성",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "optionId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductOptionItemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "옵션 항목 생성 완료"
          }
        }
      }
    },
    "/api/products/{id}/stock": {
      "patch": {
        "tags": [
          "Products"
        ],
        "summary": "재고 조정",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustStockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "재고 조정 완료"
          }
        }
      }
    },
    "/api/products/search": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "상품 검색 (관리자용)",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "storeId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "검색 결과"
          }
        }
      }
    },
...
  }
}